Handle missing extension headers
Malformed gtp-v1 packets could cause a panic if the header had the
E flag set (announcing extension headers), but was missing the
actual extension header data.

This commit introduces a test case to check for the correct error
handling and makes sure that an error is reported if an
extension header is expected but not present. Error reporting
is done by the existing 'ParseExtensionHeader' func.

Co-authored-by: Wolfgang Schäfer <wbwork@users.noreply.github.com>